reorganized top component
