+++ frontend/src/App.tsx
@@ -14,8 +14,8 @@ import PlaylistView from './views/PlaylistView';
 import Track from './views/Track';
 import Logout from './views/Logout';
 import Recent from './views/Recent';
-import Settings from './views/SettingsController';
-import Top from './views/Top';
+import Settings from './views/SettingsPage';
+import Top from './views/TopPage';
 import Compare from './views/Compare';
 import Sidebar from './views/Sidebar';
 import TrackSearch from './views/TrackSearch';
